+++ packages/nimble-components/src/select/tests/select.spec.ts
@@ -65,9 +65,22 @@ async function setup(
             <nimble-list-option value="has space">Has Space</nimble-list-option>
         </nimble-select>
     `;
-    return fixture<Select>(viewTemplate);
+    return await fixture<Select>(viewTemplate);
 }
 
+async function setupWithSpanLabels(): Promise<Fixture<Select>> {
+    const viewTemplate = html`
+        <nimble-select>
+            <nimble-list-option value="one">
+                <span>One</span>
+            </nimble-list-option>
+            <nimble-list-option value="two">
+                <span>Two</span>
+            </nimble-list-option>
+        </nimble-select>
+    `;
+    return await fixture<Select>(viewTemplate);
+}
 async function setupWithGroups(): Promise<Fixture<Select>> {
     const viewTemplate = html`
         <nimble-select>
@@ -87,7 +100,7 @@ async function setupWithGroups(): Promise<Fixture<Select>> {
             </nimble-list-option-group>
         </nimble-select>
     `;
-    return fixture<Select>(viewTemplate);
+    return await fixture<Select>(viewTemplate);
 }
 
 async function clickAndWaitForOpen(select: Select): Promise<void> {
@@ -738,6 +751,77 @@ describe('Select', () => {
         await disconnect();
     });
 
+    it('when selected option textContent is updated directly in DOM, display text is updated', async () => {
+        const { element, connect, disconnect } = await setup();
+        const pageObject = new SelectPageObject(element);
+        await connect();
+        await waitForUpdatesAsync();
+        element.value = 'two';
+        await waitForUpdatesAsync();
+        // update the textContent of the node directly to bypass FAST's textContent handling
+        element.options[1]!.childNodes[0]!.textContent = 'foo';
+        await waitForUpdatesAsync();
+
+        expect(pageObject.getDisplayText()).toBe('foo');
+
+        await disconnect();
+    });
+
+    it('when selection option textContent of child span is updated directly in DOM, display text is updated', async () => {
+        const { element, connect, disconnect } = await setupWithSpanLabels();
+        const pageObject = new SelectPageObject(element);
+        await connect();
+        await waitForUpdatesAsync();
+        element.value = 'two';
+        await waitForUpdatesAsync();
+        // update the textContent of the node directly to bypass FAST's textContent handling
+        element.options[1]!.childNodes[1]!.textContent = 'foo'; // span is at index 1 of childNodes
+        await waitForUpdatesAsync();
+
+        expect(pageObject.getDisplayText()).toBe('foo');
+
+        await disconnect();
+    });
+
+    it('when select is disconnected and reconnected, updating selected option textContent in DOM updates display text', async () => {
+        const { element, connect, disconnect } = await setup();
+        const pageObject = new SelectPageObject(element);
+        await connect();
+        await waitForUpdatesAsync();
+        element.value = 'two';
+        await waitForUpdatesAsync();
+        await disconnect();
+        await connect();
+
+        // update the textContent of the node directly to bypass FAST's textContent handling
+        element.options[1]!.childNodes[0]!.textContent = 'foo';
+        await waitForUpdatesAsync();
+
+        expect(pageObject.getDisplayText()).toBe('foo');
+
+        await disconnect();
+    });
+
+    it('when select replaces its options, updating newly selected option textContent in DOM updates display text', async () => {
+        const { element, connect, disconnect } = await setup();
+        const pageObject = new SelectPageObject(element);
+        await connect();
+        await waitForUpdatesAsync();
+        element.value = 'two';
+        await waitForUpdatesAsync();
+        await pageObject.setOptions([
+            new ListOption('one', 'one'),
+            new ListOption('two', 'two') // will keep the same value and selectedIndex, but a new option
+        ]);
+
+        element.options[1]!.childNodes[0]!.textContent = 'foo';
+        await waitForUpdatesAsync();
+
+        expect(pageObject.getDisplayText()).toBe('foo');
+
+        await disconnect();
+    });
+
     describe('with all options disabled', () => {
         async function setupAllDisabled(): Promise<Fixture<Select>> {
             const viewTemplate = html`
@@ -753,7 +837,7 @@ describe('Select', () => {
                     >
                 </nimble-select>
             `;
-            return fixture<Select>(viewTemplate);
+            return await fixture<Select>(viewTemplate);
         }
 
         it('when all options disabled, first option is selected', async () => {
@@ -782,7 +866,7 @@ describe('Select', () => {
                     >
                 </nimble-select>
             `;
-            return fixture<Select>(viewTemplate);
+            return await fixture<Select>(viewTemplate);
         }
 
         it('when all options hidden, first option is selected', async () => {
@@ -805,7 +889,7 @@ describe('Select', () => {
                         <nimble-list-option value="${x => x}">${x => x}</nimble-list-option>`)}
                 </${selectTag}>
             `;
-            return fixture<Select>(viewTemplate);
+            return await fixture<Select>(viewTemplate);
         }
 
         // Intermittent, see: https://github.com/ni/nimble/issues/2269
@@ -860,7 +944,7 @@ describe('Select', () => {
                     </${selectTag}>
                 </div>
             `;
-            return fixture<Select>(viewTemplate);
+            return await fixture<Select>(viewTemplate);
         }
 
         // Intermittent, see: https://github.com/ni/nimble/issues/2272
@@ -1614,7 +1698,6 @@ describe('Select', () => {
 
         it('clear button is visible after selecting an option', async () => {
             pageObject.clickClearButton();
-            await waitForUpdatesAsync();
             expect(pageObject.isClearButtonVisible()).toBeFalse();
             await clickAndWaitForOpen(element);
             pageObject.clickOptionWithDisplayText('Two');
@@ -1623,9 +1706,8 @@ describe('Select', () => {
             expect(pageObject.isClearButtonVisible()).toBeTrue();
         });
 
-        it('clicking clear button does not open dropdown', async () => {
+        it('clicking clear button does not open dropdown', () => {
             pageObject.clickClearButton();
-            await waitForUpdatesAsync();
             expect(element.open).toBeFalse();
         });
 
@@ -1660,9 +1742,8 @@ describe('Select', () => {
         });
 
         describe('without placeholder', () => {
-            it('after clicking clear button, display text is empty and clear button is hidden', async () => {
+            it('after clicking clear button, display text is empty and clear button is hidden', () => {
                 pageObject.clickClearButton();
-                await waitForUpdatesAsync();
 
                 expect(pageObject.getDisplayText()).toBe('');
                 expect(pageObject.isClearButtonVisible()).toBeFalse();
@@ -1694,7 +1775,6 @@ describe('Select', () => {
                 pageObject.clickOptionWithDisplayText('Two');
                 await waitForUpdatesAsync();
                 pageObject.clickClearButton();
-                await waitForUpdatesAsync();
 
                 expect(pageObject.getDisplayText()).toBe('One');
             });
@@ -1719,7 +1799,6 @@ describe('Select', () => {
                     await waitForUpdatesAsync();
                     expect(pageObject.isClearButtonVisible()).toBeTrue();
                     pageObject.clickClearButton();
-                    await waitForUpdatesAsync();
 
                     expect(pageObject.getDisplayText()).toBe('');
                 });