+++ packages/nimble-components/src/table/components/header/tests/table-header.spec.ts
@@ -6,7 +6,7 @@ import { TableColumnAlignment, TableColumnSortDirection } from '../../../types';
 import { TableHeaderPageObject } from './table-header-pageobject';
 
 async function setup(): Promise<Fixture<TableHeader>> {
-    return fixture<TableHeader>(
+    return await fixture<TableHeader>(
         html`<nimble-table-header> </nimble-table-header>`
     );
 }