+++ packages/nimble-components/src/table/tests/table-selection.spec.ts
@@ -209,7 +209,7 @@ const hierarchicalData = [
 
 // prettier-ignore
 async function setup(): Promise<Fixture<Table<SimpleTableRecord>>> {
-    return fixture<Table<SimpleTableRecord>>(
+    return await fixture<Table<SimpleTableRecord>>(
         html`<nimble-table>
             <nimble-table-column-text id="column1" field-name="stringData">stringData</nimble-table-column-text>
             <nimble-table-column-text id="column2" field-name="stringData2">stringData2</nimble-table-column-text>