+++ packages/storybook/.storybook/preview.js
@@ -105,5 +105,5 @@ configureActions({
     depth: 1
 });
 
-// Update the GUID on this line to trigger a turbosnap full rebuild: 35fca962-ed23-4476-979e-094c9a0b48f3
+// Update the GUID on this line to trigger a turbosnap full rebuild: acea85eb-3c9d-49c3-ae4f-7c12f82bc5a0
 // See https://www.chromatic.com/docs/turbosnap/#full-rebuilds